Added web KPI metrics constants and refactored Web component to utilize new hooks
- Introduced WEB_KPI_METRICS in constants.ts for better organization of KPI metrics.
- Refactored Web component to use useWebData hook for data fetching and state management.
- Updated WebKPIs component to reference new WEB_KPI_METRICS constant.
