+++ apps/stats/src/views/Stats/Web/constants.ts
@@ -1,0 +1,29 @@
+import {KpiMetric} from '@src/types/kpi';
+import {formatDuration, formatNumber, formatPercentage} from '@tryghost/shade';
+
+export const WEB_KPI_METRICS: Record<string, KpiMetric> = {
+    visits: {
+        dataKey: 'visits',
+        label: 'Visitors',
+        chartColor: 'hsl(var(--chart-blue))',
+        formatter: formatNumber
+    },
+    views: {
+        dataKey: 'pageviews',
+        label: 'Pageviews',
+        chartColor: 'hsl(var(--chart-teal))',
+        formatter: formatNumber
+    },
+    'bounce-rate': {
+        dataKey: 'bounce_rate',
+        label: 'Bounce rate',
+        chartColor: 'hsl(var(--chart-teal))',
+        formatter: formatPercentage
+    },
+    'visit-duration': {
+        dataKey: 'avg_session_sec',
+        label: 'Visit duration',
+        chartColor: 'hsl(var(--chart-teal))',
+        formatter: formatDuration
+    }
+};
